requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.2
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.1.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
orjson==3.10.18
packaging==25.0
pandas==2.3.0
pillow==11.2.1
protobuf==5.29.5
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-addons==0.23.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2024.8.30
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typeguard==2.13.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
